Refactor cucumber steps and add 2 short tests
