+++ src/tests/cucumber/features/environment.py
@@ -12,8 +12,12 @@ def check_userdata_exists(context, data_name: str):
             f"The following parameter should be defined in 'behave.ini' or in the command line when calling behave: {data_name}")
 
 def after_scenario(context, scenario):
-    # post-processing a test: clean up output files to avoid taking up all the disk space
-    if hasattr(context, "output_path") and  context.output_path != None:
-        rmtree(context.output_path)
+    # post-processing a test: remove tmp copy of study
+    if hasattr(context, "tmp_workdir") and context.tmp_workdir is not None:
+        rmtree(context.tmp_workdir)
+        return
+    # else, clean up output files to avoid taking up all the disk space
+    if hasattr(context, "output_path") and context.output_path is not None:
+        pass # rmtree(context.output_path)
     if os.path.exists(os.path.join(context.study_path, "logs")):
         rmtree(os.path.join(context.study_path, "logs"))
