+++ src/tests/cucumber/features/modeler-features/epic2/us2.8.feature
@@ -1,7 +1,7 @@
 Feature: 1 - Simple end-to-end tests to test temporal expression management
 
   Scenario: 1.1: One model with one load and two generators, two timesteps
-    Given the study path is "modeler/1_1"
+    Given the modeler study path is "modeler/1_1"
     When I run antares modeler
     Then the simulation succeeds
     And the objective value is 320
@@ -11,7 +11,7 @@ Feature: 1 - Simple end-to-end tests to test temporal expression management
       | node1     | gen2_p   | 0-1      | 20    |
 
   Scenario: 1.2: One model with one load and two generators, two timesteps ; mix of constant parameters and timeseries
-    Given the study path is "modeler/1_2"
+    Given the modeler study path is "modeler/1_2"
     When I run antares modeler
     Then the simulation succeeds
     And the objective value is 320
@@ -21,13 +21,13 @@ Feature: 1 - Simple end-to-end tests to test temporal expression management
       | node1     | gen2_p   | 0-1      | 20    |
 
   Scenario: 1.3: One model with one load and two generators, one timestep ; fails because of timeseries of different lengths
-    Given the study path is "modeler/1_3"
+    Given the modeler study path is "modeler/1_3"
     When I run antares modeler
     Then the simulation fails
 
 
   Scenario: 1.4: test system with one load (100 MW) and two generators in the same node (with Pmax_1 = 80 MW & Pmax_2 = 200 MW) ; gen1 has a fixed switch on cost. Two timesteps.
-    Given the study path is "modeler/1_4"
+    Given the modeler study path is "modeler/1_4"
     When I run antares modeler
     Then the simulation succeeds
     And the objective value is 765
@@ -42,12 +42,12 @@ Feature: 1 - Simple end-to-end tests to test temporal expression management
 
 
   Scenario: 1.5: One model with one load and two generators, one timestep ; fails because of timeseries given for parameter with time_dependant = false
-    Given the study path is "modeler/1_5"
+    Given the modeler study path is "modeler/1_5"
     When I run antares modeler
     Then the simulation fails
 
   Scenario: 1.6: One model with one load and two generators, 100 timesteps
-    Given the study path is "modeler/1_6"
+    Given the modeler study path is "modeler/1_6"
     When I run antares modeler
     Then the simulation succeeds
     And the objective value is 16000
