+++ src/tests/cucumber/features/steps/common_steps/modeler_output_handler.py
@@ -2,8 +2,6 @@
 
 import os
 import pandas as pd
-import configparser
-from enum import Enum
 
 
 class modeler_output_handler:
